refactor: create product utils, renamed product interfaces
