+++ Exercices/grocery-store/src/ProductTableRow.tsx
@@ -1,5 +1,5 @@
 import React, { FunctionComponent } from 'react'
-import { Product } from './interfaces/dbInterfaces'
+import { Product } from './interfaces/products'
 
 interface ProductTableRowProps {
   product: Product
